+++ bulldozer/utils/bulldozer_logger.py
@@ -31,8 +31,6 @@ import psutil
 import multiprocessing
 import logging
 import logging.config
-from git import Repo
-from git.exc import InvalidGitRepositoryError
 from bulldozer._version import __version__
 
 class BulldozerLogger:
@@ -115,19 +113,7 @@ class BulldozerLogger:
             This method store the environment state in the logfile.
         """
         info={}
-        try:
-            # Git info
-            try :
-                repo = Repo(search_parent_directories=True)
-                info['commit_sha'] = repo.head.object.hexsha
-                info['branch'] = repo.active_branch
-            except InvalidGitRepositoryError as e:
-                info['commit_sha'] = "No git repo found ({})".format(e)
-                info['branch'] = "No git repo found ({})".format(e)
-            except:
-                info['commit_sha'] = "unknown"
-                info['branch'] = "unknown"  
-    
+        try:  
             # Node info
             try:
                 info['user'] = getpass.getuser()
@@ -147,11 +133,11 @@ class BulldozerLogger:
             info['os_version'] = platform.version()
             
             # Message format
-            init = ("\n"+"#"*17+"\n#   BULLDOZER   #\n"+"#"*17+"\n# <Bulldozer info>\n#\t- version: {}"+"\n#\n# <Git info>\n#\t- branch: {}\n#\t- commit SHA: {}"
+            init = ("\n"+"#"*17+"\n#   BULLDOZER   #\n"+"#"*17+"\n# <Bulldozer info>\n#\t- version: {}"+
                     "\n#\n# <Node info>\n#\t - user: {}\n#\t - node: {}\n#\t - processor: {}\n#\t - CPU count: {}\n#\t - RAM: {}"
                     "\n#\n# <OS info>\n#\t - system: {}\n#\t - release: {}\n#\t - version: {}\n"
-                    +"#"*17).format(__version__, info['branch'], info['commit_sha'], info['user'], info['node'], 
-                                    info['processor'], info['cpu_count'], info['ram'], info['system'], info['release'], info['os_version'])
+                    +"#"*17).format(__version__, info['user'], info['node'], info['processor'], info['cpu_count'], info['ram'], 
+                    info['system'], info['release'], info['os_version'])
             BulldozerLogger.log(init, logging.DEBUG)
 
         except Exception as e: